Rename package starMap.randomEvent to starMap.event
Shorter, generic and more JLS compliant name.
