+++ src/main/java/org/openRealmOfStars/gui/mapPanel/PopupPanel.java
@@ -27,8 +27,8 @@ import org.openRealmOfStars.gui.util.GuiFonts;
 import org.openRealmOfStars.gui.util.GuiStatics;
 import org.openRealmOfStars.mapTiles.anomaly.SpaceAnomaly;
 import org.openRealmOfStars.player.combat.Combat;
+import org.openRealmOfStars.starMap.event.RandomEvent;
 import org.openRealmOfStars.starMap.newsCorp.ImageInstruction;
-import org.openRealmOfStars.starMap.randomEvent.RandomEvent;
 
 
 /**
