+++ src/main/java/org/openRealmOfStars/starMap/newsCorp/NewsFactory.java
@@ -29,6 +29,9 @@ import org.openRealmOfStars.starMap.Coordinate;
 import org.openRealmOfStars.starMap.CulturePower;
 import org.openRealmOfStars.starMap.StarMap;
 import org.openRealmOfStars.starMap.StarMapUtilities;
+import org.openRealmOfStars.starMap.event.BadRandomType;
+import org.openRealmOfStars.starMap.event.GoodRandomType;
+import org.openRealmOfStars.starMap.event.RandomEvent;
 import org.openRealmOfStars.starMap.history.event.DiplomaticEvent;
 import org.openRealmOfStars.starMap.history.event.Event;
 import org.openRealmOfStars.starMap.history.event.LeaderEvent;
@@ -37,9 +40,6 @@ import org.openRealmOfStars.starMap.newsCorp.scoreBoard.ScoreBoard;
 import org.openRealmOfStars.starMap.planet.Planet;
 import org.openRealmOfStars.starMap.planet.construction.Building;
 import org.openRealmOfStars.starMap.planet.enums.PlanetTypes;
-import org.openRealmOfStars.starMap.randomEvent.BadRandomType;
-import org.openRealmOfStars.starMap.randomEvent.GoodRandomType;
-import org.openRealmOfStars.starMap.randomEvent.RandomEvent;
 import org.openRealmOfStars.starMap.vote.Vote;
 import org.openRealmOfStars.starMap.vote.VotingType;
 import org.openRealmOfStars.starMap.vote.sports.Athlete;
