+++ src/test/java/org/openRealmOfStars/starMap/newsCorp/NewsFactoryTest.java
@@ -32,14 +32,14 @@ import org.openRealmOfStars.player.fleet.Fleet;
 import org.openRealmOfStars.player.fleet.FleetList;
 import org.openRealmOfStars.player.race.SpaceRace;
 import org.openRealmOfStars.starMap.StarMap;
+import org.openRealmOfStars.starMap.event.BadRandomType;
+import org.openRealmOfStars.starMap.event.GoodRandomType;
+import org.openRealmOfStars.starMap.event.RandomEvent;
 import org.openRealmOfStars.starMap.history.History;
 import org.openRealmOfStars.starMap.planet.Planet;
 import org.openRealmOfStars.starMap.planet.construction.Building;
 import org.openRealmOfStars.starMap.planet.enums.PlanetTypes;
 import org.openRealmOfStars.starMap.planet.enums.RadiationType;
-import org.openRealmOfStars.starMap.randomEvent.BadRandomType;
-import org.openRealmOfStars.starMap.randomEvent.GoodRandomType;
-import org.openRealmOfStars.starMap.randomEvent.RandomEvent;
 import org.openRealmOfStars.starMap.vote.Vote;
 import org.openRealmOfStars.starMap.vote.Votes;
 import org.openRealmOfStars.starMap.vote.VotingType;
